Added new test class, exception class

+++ src/main/java/com/cleancode/service/FlightService.java
@@ -1,5 +1,7 @@
 package com.cleancode.service;
 
+import com.cleancode.FlightNotFoundException;
+import com.cleancode.InvalidInputException;
 import com.cleancode.model.Flight;
 import com.cleancode.repository.FlightRepository;
 import org.springframework.stereotype.Service;
@@ -23,21 +25,29 @@ public class FlightService implements IFlightService{
 
     @Override
     public Flight findFlightById(Long id) {
-        return repository.findFlightById(id);
+        if (id > 0) return repository.findFlightById(id);
+        else throw new InvalidInputException("Your input is invalid.");
     }
 
     @Override
     public Optional<Flight> findById(Long id) {
-        return repository.findById(id);
+        Optional<Flight> result = repository.findById(id);
+        if(id > 0 && result.isPresent()) return result;
+        else if(id > 0) throw new FlightNotFoundException("Flight not found");
+        else throw new InvalidInputException("Your input is invalid.");
     }
 
     @Override
     public void deleteFlightById(Long id) {
-        repository.deleteById(id);
+        if (repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+        }
+        else throw new FlightNotFoundException("Flight not found");
     }
 
     @Override
     public Flight create(Flight newFlight) {
-        return repository.save(newFlight);
+        if(repository.findById(newFlight.getId()).isPresent()) throw new InvalidInputException("The flight already exists.");
+        else return repository.save(newFlight);
     }
 }
